// src/types.ts
export interface Activity {
  [key: string]: unknown;
  id: number;
  type: string;
  subject: string;
  done: boolean;
  due_date: string;
  due_time: string | null;
  duration: string | null;
  deal_id: number | null;
  person_id: number | null;
  org_id: number | null;
  user_id: number;
  note: string | null;
  add_time: string;
  update_time: string;
  marked_as_done_time: string | null;
  created_by_user_id: number;
  owner_name: string;
  deal_title: string | null;
  person_name: string | null;
  org_name: string | null;
}